refactor(cmp/util/types): Add new StopData type.
